Cleaned version, small xpm are removed, we use only the palette in Pixmap rainbow

+++ read_temperature.h
@@ -1,6 +1,6 @@
 #ifndef READTEMPERATURE
 #define READTEMPERATURE
 
-int read_temperature(char*);
+float read_temperature(char*);
 
 #endif
